practica-mqtt/requirements.txt
numpy==1.22.4
paho-mqtt==1.6.1
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
simplejson==3.17.6
six==1.16.0
